+++ docassemble_base/docassemble/base/core.py
@@ -4280,9 +4280,9 @@ class DATemplate(DAObject):
         return str(self.content)
     def __str__(self):
         if docassemble.base.functions.this_thread.evaluation_context == 'docx':
-            #return str(self.content)
-            #return str(docassemble.base.filter.docx_template_filter(self.content))
-            return str(docassemble.base.file_docx.markdown_to_docx(self.content, docassemble.base.functions.this_thread.current_question, docassemble.base.functions.this_thread.misc.get('docx_template', None)))
+            content = self.content
+            content = re.sub(r'\\_', r'\\\\_', content)
+            return str(docassemble.base.file_docx.markdown_to_docx(content, docassemble.base.functions.this_thread.current_question, docassemble.base.functions.this_thread.misc.get('docx_template', None)))
         return(str(self.content))
 
 def table_safe(text):
@@ -4471,8 +4471,6 @@ class DALazyTemplate(DAObject):
         if docassemble.base.functions.this_thread.evaluation_context == 'docx':
             content = self.content
             content = re.sub(r'\\_', r'\\\\_', content)
-            #return str(self.content)
-            #return str(docassemble.base.filter.docx_template_filter(self.content))
             return str(docassemble.base.file_docx.markdown_to_docx(content, docassemble.base.functions.this_thread.current_question, docassemble.base.functions.this_thread.misc.get('docx_template', None)))
         return(str(self.content))